requirements.txt
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-JWT-Extended==4.4.2
Flask-SQLAlchemy==2.5.1
newrelic==7.14.0.177
gunicorn==20.1.0
psycopg2==2.9.3